chore: broke down files and added jsdocs to StepNode and StepsLinkedList
