+++ apps/meteor/client/components/Wizard/lib/StepsLinkedList.ts
@@ -1,69 +1,39 @@
 import { Emitter } from '@rocket.chat/emitter';
 
-export type StepMetadata = {
-	id: string;
-	title: string;
-	onNext?(): void;
-	onPrev?(): void;
-};
+import type { StepMetadata } from './StepNode';
+import StepNode from './StepNode';
 
-export class StepNode {
-	public value: StepMetadata;
-
-	public next: StepNode | null = null;
-
-	public prev: StepNode | null = null;
-
-	public disabled = true;
-
-	constructor(value: StepMetadata, disabled = true) {
-		this.value = value;
-		this.disabled = disabled;
-	}
-
-	enable() {
-		this.disabled = false;
-	}
-
-	disable() {
-		this.disabled = true;
-	}
-
-	get id() {
-		return this.value.id;
-	}
-
-	get title() {
-		return this.value.title;
-	}
-
-	onNext() {
-		this.value.onNext?.();
-	}
-
-	onPrev() {
-		this.value.onPrev?.();
-	}
-}
-
-export class StepsLinkedList extends Emitter<{ stateChanged: undefined }> {
+/**
+ * A doubly linked list implementation to manage the state of wizard steps.
+ * It extends Emitter to notify about state changes.
+ */
+class StepsLinkedList extends Emitter<{ stateChanged: undefined }> {
 	public head: StepNode | null = null;
 
 	public tail: StepNode | null = null;
 
 	private stepNodeMap: Map<string | number, StepNode> = new Map();
 
-	constructor(steps: StepMetadata[]) {
+	/**
+	 * Creates an instance of StepsLinkedList.
+	 * @param steps - An array of step metadata to initialize the list with.
+	 */
+	constructor(steps: readonly StepMetadata[]) {
 		super();
 		steps.forEach((step) => this.append(step));
 	}
 
+	/**
+	 * Appends a new step to the end of the list or updates an existing one.
+	 * @param value - The metadata for the step to append or update.
+	 * @returns The created or updated StepNode.
+	 */
 	public append(value: StepMetadata): StepNode {
 		const existingStep = this.stepNodeMap.get(value.id);
 
 		if (existingStep) {
 			existingStep.value = value;
-			this.emit('stateChanged');
+			this.notifyChanges();
 			return existingStep;
 		}
 
@@ -82,6 +52,11 @@ export class StepsLinkedList extends Emitter<{ stateChanged: undefined }> {
 		return newNode;
 	}
 
+	/**
+	 * Removes a step from the list by its ID.
+	 * @param id - The ID of the step to remove.
+	 * @returns The removed StepNode, or null if not found.
+	 */
 	public remove(id: string | number): StepNode | null {
 		const nodeToRemove = this.stepNodeMap.get(id);
 
@@ -109,30 +84,55 @@ export class StepsLinkedList extends Emitter<{ stateChanged: undefined }> {
 		return nodeToRemove;
 	}
 
+	/**
+	 * Enables a specific step.
+	 * @param step - The StepNode to enable.
+	 */
 	public enableStep(step: StepNode) {
 		step.enable();
 		this.notifyChanges();
 	}
 
+	/**
+	 * Disables a specific step.
+	 * @param step - The StepNode to disable.
+	 */
 	public disableStep(step: StepNode) {
 		step.disable();
 		this.notifyChanges();
 	}
 
+	/**
+	 * Emits a 'stateChanged' event to notify listeners of changes.
+	 */
 	private notifyChanges() {
 		this.emit('stateChanged');
 	}
 
+	/**
+	 * Retrieves a step by its ID.
+	 * @param id - The ID of the step to retrieve.
+	 * @returns The StepNode if found, otherwise null.
+	 */
 	public get(id: string | number): StepNode | null {
 		const node = this.stepNodeMap.get(id);
 		return node || null;
 	}
 
+	/**
+	 * Converts the linked list to an array of StepNodes.
+	 * @returns An array of all StepNodes in the list.
+	 */
 	public toArray(): StepNode[] {
 		return Array.from(this.stepNodeMap.values());
 	}
 
+	/**
+	 * Gets the total number of steps in the list.
+	 */
 	get size(): number {
 		return this.stepNodeMap.size;
 	}
 }
+
+export default StepsLinkedList;
